+++ src/Event/AfterFilenameSanitizationEvent.php
@@ -1,0 +1,45 @@
+<?php
+
+/*
+ * Copyright (c) 2019 Heimrich & Hannot GmbH
+ *
+ * @license LGPL-3.0-or-later
+ */
+
+namespace HeimrichHannot\FilenameSanitizerBundle\Event;
+
+use Symfony\Component\EventDispatcher\Event;
+
+class AfterFilenameSanitizationEvent extends Event
+{
+    const NAME = 'huh.filename_sanitizer.event.after_filename_sanitization_event';
+
+    /**
+     * @var string
+     */
+    protected $filename;
+
+    /**
+     * @param string $filename
+     */
+    public function __construct(string $filename)
+    {
+        $this->filename = $filename;
+    }
+
+    /**
+     * @return string
+     */
+    public function getFilename(): string
+    {
+        return $this->filename;
+    }
+
+    /**
+     * @param string $filename
+     */
+    public function setFilename(string $filename): void
+    {
+        $this->filename = $filename;
+    }
+}