+++ src/Event/AfterStringSanitizationEvent.php
@@ -1,0 +1,45 @@
+<?php
+
+/*
+ * Copyright (c) 2019 Heimrich & Hannot GmbH
+ *
+ * @license LGPL-3.0-or-later
+ */
+
+namespace HeimrichHannot\FilenameSanitizerBundle\Event;
+
+use Symfony\Component\EventDispatcher\Event;
+
+class AfterStringSanitizationEvent extends Event
+{
+    const NAME = 'huh.filename_sanitizer.event.after_string_sanitization_event';
+
+    /**
+     * @var string
+     */
+    protected $string;
+
+    /**
+     * @param string $string
+     */
+    public function __construct(string $string)
+    {
+        $this->string = $string;
+    }
+
+    /**
+     * @return string
+     */
+    public function getString(): string
+    {
+        return $this->string;
+    }
+
+    /**
+     * @param string $string
+     */
+    public function setString(string $string): void
+    {
+        $this->string = $string;
+    }
+}