+++ src/Resources/contao/config/config.php
@@ -5,5 +5,5 @@
  */
 $GLOBALS['TL_HOOKS']['postUpload']['sanitizeFilename'] = ['huh.filename_sanitizer.event_listener.hook_listener', 'sanitizeFilenames'];
 
-$settingsService = System::getContainer()->get('huh.filename_sanitizer.data_container.settings');
+$settingsService = System::getContainer()->get('huh.filename_sanitizer.data_container.settings_container');
 $settingsService->setDefaults();