+++ src/Resources/contao/dca/tl_files.php
@@ -2,4 +2,4 @@
 
 $dca = &$GLOBALS['TL_DCA']['tl_files'];
 
-$dca['config']['onsubmit_callback']['sanitizeFilename'] = ['huh.filename_sanitizer.data_container.files', 'sanitizeFilename'];
+$dca['config']['onsubmit_callback']['sanitizeFilename'] = ['huh.filename_sanitizer.data_container.files_container', 'sanitizeFilename'];