+++ src/Resources/contao/languages/de/default.php
@@ -1,3 +1,0 @@
-<?php
-
-//$lang = &$GLOBALS['TL_LANG']['MSC']['listBundle'];