+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.2.0] - 2019-03-18
+
+### Added
+- character replacement feature
+- English localization
+- default character replacement for German umlauts (can be deactivated in global Contao settings)
+- multiple symfony events
+- image for readme
+
+### Changed
+- internal service naming
+
 ## [1.1.2] - 2019-03-06
 
 ### Fixed